refactor: 테이블 그룹 해제시 Event사용 대신 TableService에서 처리
