+++ src/main/java/kitchenpos/ordertable/application/TableService.java
@@ -15,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
@@ -64,6 +65,25 @@ public class TableService {
 
     private void validateOrderStatusToChangeEmpty(Long orderTableId) {
         if(orderRepository.existsByOrderTableIdAndOrderStatusIn(orderTableId,
+                Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalOrderException(
+                    String.format(ERROR_ORDER_INVALID_STATUS, OrderStatus.COOKING + " " + OrderStatus.MEAL)
+            );
+        }
+    }
+
+    @Transactional
+    public void ungroupTableByTableGroupId(Long tableGroupId){
+        List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(tableGroupId);
+        validateOrderTablesToUngroup(orderTables);
+        orderTables.forEach(orderTable -> orderTable.removeTableGroup());
+    }
+
+    private void validateOrderTablesToUngroup(List<OrderTable> orderTables) {
+        final List<Long> orderTableIds = orderTables.stream()
+                .map(OrderTable::getId)
+                .collect(Collectors.toList());
+        if(orderRepository.existsByOrderTableIdInAndOrderStatusIn(orderTableIds,
                 Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
             throw new IllegalOrderException(
                     String.format(ERROR_ORDER_INVALID_STATUS, OrderStatus.COOKING + " " + OrderStatus.MEAL)
