--- src/main/java/kitchenpos/ordertable/event/handler/UngroupTableEventHandler.java
@@ -1,49 +1,0 @@
-package kitchenpos.ordertable.event.handler;
-
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.exception.IllegalOrderException;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTableRepository;
-import kitchenpos.ordertable.event.UngroupTableEvent;
-import org.springframework.context.event.EventListener;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Component
-public class UngroupTableEventHandler {
-    private final OrderTableRepository orderTableRepository;
-    private final OrderRepository orderRepository;
-
-    public static final String ERROR_ORDER_INVALID_STATUS = "주문의 상태는 %s일 수 없습니다.";
-
-    public UngroupTableEventHandler(OrderTableRepository orderTableRepository, OrderRepository orderRepository) {
-        this.orderTableRepository = orderTableRepository;
-        this.orderRepository = orderRepository;
-    }
-
-    @Transactional
-    @EventListener
-    public void handle(UngroupTableEvent event) {
-        List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(event.getTableGroupId());
-        validateOrderTablesToUngroup(orderTables);
-
-        orderTables.forEach(orderTable -> orderTable.removeTableGroup());
-    }
-
-    private void validateOrderTablesToUngroup(List<OrderTable> orderTables) {
-        final List<Long> orderTableIds = orderTables.stream()
-                .map(OrderTable::getId)
-                .collect(Collectors.toList());
-        if(orderRepository.existsByOrderTableIdInAndOrderStatusIn(orderTableIds,
-                Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalOrderException(
-                    String.format(ERROR_ORDER_INVALID_STATUS, OrderStatus.COOKING + " " + OrderStatus.MEAL)
-            );
-        }
-    }
-}
